lsm6dsr: examples updated
