+++ lsm6dsr_STdC/example/lsm6dsr_fifo_pedo.c
@@ -7,7 +7,7 @@
  ******************************************************************************
  * @attention
  *
- * <h2><center>&copy; Copyright (c) 2019 STMicroelectronics.
+ * <h2><center>&copy; Copyright (c) 2021 STMicroelectronics.
  * All rights reserved.</center></h2>
  *
  * This software component is licensed by ST under BSD 3-Clause license,
@@ -22,18 +22,20 @@
  * This example was developed using the following STMicroelectronics
  * evaluation boards:
  *
- * - STEVAL_MKI109V3
- * - NUCLEO_F411RE + X_NUCLEO_IKS01A2
- *
- * and STM32CubeMX tool with STM32CubeF4 MCU Package
+ * - STEVAL_MKI109V3 + STEVAL-MKI194V1
+ * - NUCLEO_F411RE + X_STEVAL-MKI194V1
+ * - DISCOVERY_SPC584B + STEVAL-MKI194V1
  *
  * Used interfaces:
  *
  * STEVAL_MKI109V3    - Host side:   USB (Virtual COM)
  *                    - Sensor side: SPI(Default) / I2C(supported)
  *
- * NUCLEO_STM32F411RE + X_NUCLEO_IKS01A2 - Host side: UART(COM) to USB bridge
- *                                       - I2C(Default) / SPI(N/A)
+ * NUCLEO_STM32F411RE - Host side: UART(COM) to USB bridge
+ *                    - I2C(Default) / SPI(supported)
+ *
+ * DISCOVERY_SPC584B  - Host side: UART(COM) to USB bridge
+ *                    - Sensor side: I2C(Default) / SPI(supported)
  *
  * If you need to run this example on a different hardware platform a
  * modification of the functions: `platform_write`, `platform_read`,
@@ -47,37 +49,56 @@
  * If a different hardware is used please comment all
  * following target board and redefine yours.
  */
-//#define STEVAL_MKI109V3
-#define NUCLEO_F411RE_X_NUCLEO_IKS01A2
+
+//#define STEVAL_MKI109V3  /* little endian */
+//#define NUCLEO_F411RE    /* little endian */
+//#define SPC584B_DIS      /* big endian */
+
+/* ATTENTION: By default the driver is little endian. If you need switch
+ *            to big endian please see "Endianness definitions" in the
+ *            header file of the driver (_reg.h).
+ */
 
 #if defined(STEVAL_MKI109V3)
 /* MKI109V3: Define communication interface */
 #define SENSOR_BUS hspi2
-
 /* MKI109V3: Vdd and Vddio power supply values */
 #define PWM_3V3 915
 
-#elif defined(NUCLEO_F411RE_X_NUCLEO_IKS01A2)
-/* NUCLEO_F411RE_X_NUCLEO_IKS01A2: Define communication interface */
+#elif defined(NUCLEO_F411RE)
+/* NUCLEO_F411RE: Define communication interface */
 #define SENSOR_BUS hi2c1
+
+#elif defined(SPC584B_DIS)
+/* DISCOVERY_SPC584B: Define communication interface */
+#define SENSOR_BUS I2CD1
 
 #endif
 
 /* Includes ------------------------------------------------------------------*/
 #include <string.h>
 #include <stdio.h>
+#include "lsm6dsr_reg.h"
+
+#if defined(NUCLEO_F411RE)
 #include "stm32f4xx_hal.h"
-#include <lsm6dsr_reg.h>
+#include "usart.h"
 #include "gpio.h"
 #include "i2c.h"
-#if defined(STEVAL_MKI109V3)
+
+#elif defined(STEVAL_MKI109V3)
+#include "stm32f4xx_hal.h"
 #include "usbd_cdc_if.h"
+#include "gpio.h"
 #include "spi.h"
-#elif defined(NUCLEO_F411RE_X_NUCLEO_IKS01A2)
-#include "usart.h"
+#include "tim.h"
+
+#elif defined(SPC584B_DIS)
+#include "components.h"
 #endif
 
 /* Private macro -------------------------------------------------------------*/
+#define    BOOT_TIME            10 //ms
 
 /* Private variables ---------------------------------------------------------*/
 typedef union {
@@ -111,70 +132,50 @@ static int32_t platform_write(void *handle, uint8_t reg,
 static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
                              uint16_t len);
 static void tx_com( uint8_t *tx_buffer, uint16_t len );
+static void platform_delay(uint32_t ms);
 static void platform_init(void);
 
 /* Main Example --------------------------------------------------------------*/
-void lsm6dsr_fifo_pedo_simple(void)
+void lsm6dsr_fifo_pedo(void)
 {
   stmdev_ctx_t ag_ctx;
-  /*
-   * Uncomment to configure INT 1
-   */
+  /* Uncomment to configure INT 1 */
   //lsm6dsr_pin_int1_route_t int1_route;
-  /*
-   * Uncomment to configure INT 2
-   */
+  /* Uncomment to configure INT 2 */
   //lsm6dsr_pin_int2_route_t int2_route;
   ag_ctx.write_reg = platform_write;
   ag_ctx.read_reg = platform_read;
-  ag_ctx.handle = &hi2c1;
-  /*
-   * Init test platform
-   */
+  ag_ctx.handle = &SENSOR_BUS;
+  /* Init test platform */
   platform_init();
-  /*
-   *  Check device ID
-   */
+  /* Wait sensor boot time */
+  platform_delay(BOOT_TIME);
+  /* Check device ID */
   lsm6dsr_device_id_get(&ag_ctx, &whoamI);
 
   if (whoamI != LSM6DSR_ID)
     while (1);
 
-  /*
-   *  Restore default configuration
-   */
+  /* Restore default configuration */
   lsm6dsr_reset_set(&ag_ctx, PROPERTY_ENABLE);
 
   do {
     lsm6dsr_reset_get(&ag_ctx, &rst);
   } while (rst);
 
-  /*
-   * Disable I3C interface
-   */
+  /* Disable I3C interface */
   lsm6dsr_i3c_disable_set(&ag_ctx, LSM6DSR_I3C_DISABLE);
-  /*
-   * Set XL full scale
-   */
+  /* Set XL full scale */
   lsm6dsr_xl_full_scale_set(&ag_ctx, LSM6DSR_2g);
-  /*
-   *  Enable Block Data Update
-   */
+  /* Enable Block Data Update */
   lsm6dsr_block_data_update_set(&ag_ctx, PROPERTY_ENABLE);
-  /*
-   * Set FIFO mode to Stream mode (aka Continuous Mode)
-   */
+  /* Set FIFO mode to Stream mode (aka Continuous Mode) */
   lsm6dsr_fifo_mode_set(&ag_ctx, LSM6DSR_STREAM_MODE);
-  /*
-   * Enable latched interrupt notification.
-   */
+  /* Enable latched interrupt notification. */
   lsm6dsr_int_notification_set(&ag_ctx, LSM6DSR_ALL_INT_LATCHED);
-  /*
-    * Enable drdy 75 μs pulse: uncomment if interrupt must be pulsed.
-    */
+  /* Enable drdy 75 μs pulse: uncomment if interrupt must be pulsed. */
   //lsm6dsr_data_ready_mode_set(&ag_ctx, LSM6DSR_DRDY_PULSED);
-  /*
-   * FIFO watermark interrupt routed on INT1 pin
+  /* FIFO watermark interrupt routed on INT1 pin
    *
    * Remember that INT1 pin is used by sensor to switch in I3C mode
    * Uncomment to configure INT 1
@@ -193,35 +194,25 @@ void lsm6dsr_fifo_pedo_simple(void)
    * Enable HW Timestamp
    */
   lsm6dsr_timestamp_set(&ag_ctx, PROPERTY_ENABLE);
-  /*
-   * Enable pedometer
-   */
+  /* Enable pedometer */
   lsm6dsr_pedo_sens_set(&ag_ctx, PROPERTY_ENABLE);
   lsm6dsr_fifo_pedo_batch_set(&ag_ctx, PROPERTY_ENABLE);
   lsm6dsr_steps_reset(&ag_ctx);
-  /*
-   * Enable I2C Master
-   */
+  /* Enable I2C Master */
   lsm6dsr_sh_master_set(&ag_ctx, PROPERTY_ENABLE);
-  /*
-   * Set Output Data Rate
-   */
+  /* Set Output Data Rate */
   lsm6dsr_xl_data_rate_set(&ag_ctx, LSM6DSR_XL_ODR_26Hz);
 
   while (1) {
     uint16_t num = 0;
     lsm6dsr_fifo_tag_t reg_tag;
     pedo_count_sample_t pedo_sample;
-    /*
-     * Read FIFO samples number
-     */
+    /* Read FIFO samples number */
     lsm6dsr_fifo_data_level_get(&ag_ctx, &num);
 
     if (num > 0) {
       while (num--) {
-        /*
-         * Read FIFO tag
-         */
+        /* Read FIFO tag */
         lsm6dsr_fifo_sensor_tag_get(&ag_ctx, &reg_tag);
 
         switch (reg_tag) {
@@ -255,20 +246,16 @@ static int32_t platform_write(void *handle, uint8_t reg,
                               uint8_t *bufp,
                               uint16_t len)
 {
-  if (handle == &hi2c1) {
-    HAL_I2C_Mem_Write(handle, LSM6DSR_I2C_ADD_L, reg,
-                      I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
-  }
-
-#ifdef STEVAL_MKI109V3
-
-  else if (handle == &hspi2) {
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
-    HAL_SPI_Transmit(handle, &reg, 1, 1000);
-    HAL_SPI_Transmit(handle, bufp, len, 1000);
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
-  }
-
+#if defined(NUCLEO_F411RE)
+  HAL_I2C_Mem_Write(handle, LSM6DSR_I2C_ADD_L, reg,
+                    I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
+#elif defined(STEVAL_MKI109V3)
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
+  HAL_SPI_Transmit(handle, &reg, 1, 1000);
+  HAL_SPI_Transmit(handle, bufp, len, 1000);
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
+#elif defined(SPC584B_DIS)
+  i2c_lld_write(handle,  LSM6DSR_I2C_ADD_L & 0xFE, reg, bufp, len);
 #endif
   return 0;
 }
@@ -286,28 +273,23 @@ static int32_t platform_write(void *handle, uint8_t reg,
 static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
                              uint16_t len)
 {
-  if (handle == &hi2c1) {
-    HAL_I2C_Mem_Read(handle, LSM6DSR_I2C_ADD_L, reg,
-                     I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
-  }
-
-#ifdef STEVAL_MKI109V3
-
-  else if (handle == &hspi2) {
-    /* Read command */
-    reg |= 0x80;
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
-    HAL_SPI_Transmit(handle, &reg, 1, 1000);
-    HAL_SPI_Receive(handle, bufp, len, 1000);
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
-  }
-
+#if defined(NUCLEO_F411RE)
+  HAL_I2C_Mem_Read(handle, LSM6DSR_I2C_ADD_L, reg,
+                   I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
+#elif defined(STEVAL_MKI109V3)
+  reg |= 0x80;
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
+  HAL_SPI_Transmit(handle, &reg, 1, 1000);
+  HAL_SPI_Receive(handle, bufp, len, 1000);
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
+#elif defined(SPC584B_DIS)
+  i2c_lld_read(handle, LSM6DSR_I2C_ADD_L & 0xFE, reg, bufp, len);
 #endif
   return 0;
 }
 
 /*
- * @brief  Write generic device register (platform dependent)
+ * @brief  Send buffer to console (platform dependent)
  *
  * @param  tx_buffer     buffer to transmit
  * @param  len           number of byte to send
@@ -315,11 +297,27 @@ static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
  */
 static void tx_com(uint8_t *tx_buffer, uint16_t len)
 {
-#ifdef NUCLEO_F411RE_X_NUCLEO_IKS01A2
+#if defined(NUCLEO_F411RE)
   HAL_UART_Transmit(&huart2, tx_buffer, len, 1000);
-#endif
-#ifdef STEVAL_MKI109V3
+#elif defined(STEVAL_MKI109V3)
   CDC_Transmit_FS(tx_buffer, len);
+#elif defined(SPC584B_DIS)
+  sd_lld_write(&SD2, tx_buffer, len);
+#endif
+}
+
+/*
+ * @brief  platform specific delay (platform dependent)
+ *
+ * @param  ms        delay in ms
+ *
+ */
+static void platform_delay(uint32_t ms)
+{
+#if defined(NUCLEO_F411RE) | defined(STEVAL_MKI109V3)
+  HAL_Delay(ms);
+#elif defined(SPC584B_DIS)
+  osalThreadDelayMilliseconds(ms);
 #endif
 }
 
@@ -328,9 +326,11 @@ static void tx_com(uint8_t *tx_buffer, uint16_t len)
  */
 static void platform_init(void)
 {
-#ifdef STEVAL_MKI109V3
+#if defined(STEVAL_MKI109V3)
   TIM3->CCR1 = PWM_3V3;
   TIM3->CCR2 = PWM_3V3;
+  HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_1);
+  HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_2);
   HAL_Delay(1000);
 #endif
 }
