+++ lsm6dsr_STdC/example/lsm6dsr_tilt.c
@@ -1,14 +1,13 @@
 /*
  ******************************************************************************
- * @file    orientation_6d_4d.c
+ * @file    tilt.c
  * @author  Sensors Software Solution Team
- * @brief   This file show the simplest way to detect orientation 6D/4D event
- *      from sensor.
+ * @brief   This file show how to detect tilt event from sensor.
  *
  ******************************************************************************
  * @attention
  *
- * <h2><center>&copy; Copyright (c) 2019 STMicroelectronics.
+ * <h2><center>&copy; Copyright (c) 2021 STMicroelectronics.
  * All rights reserved.</center></h2>
  *
  * This software component is licensed by ST under BSD 3-Clause license,
@@ -23,18 +22,20 @@
  * This example was developed using the following STMicroelectronics
  * evaluation boards:
  *
- * - STEVAL_MKI109V3
- * - NUCLEO_F411RE + X_NUCLEO_IKS01A2
- *
- * and STM32CubeMX tool with STM32CubeF4 MCU Package
+ * - STEVAL_MKI109V3 + STEVAL-MKI194V1
+ * - NUCLEO_F411RE + X_STEVAL-MKI194V1
+ * - DISCOVERY_SPC584B + STEVAL-MKI194V1
  *
  * Used interfaces:
  *
  * STEVAL_MKI109V3    - Host side:   USB (Virtual COM)
  *                    - Sensor side: SPI(Default) / I2C(supported)
  *
- * NUCLEO_STM32F411RE + X_NUCLEO_IKS01A2 - Host side: UART(COM) to USB bridge
- *                                       - I2C(Default) / SPI(N/A)
+ * NUCLEO_STM32F411RE - Host side: UART(COM) to USB bridge
+ *                    - I2C(Default) / SPI(supported)
+ *
+ * DISCOVERY_SPC584B  - Host side: UART(COM) to USB bridge
+ *                    - Sensor side: I2C(Default) / SPI(supported)
  *
  * If you need to run this example on a different hardware platform a
  * modification of the functions: `platform_write`, `platform_read`,
@@ -48,37 +49,56 @@
  * If a different hardware is used please comment all
  * following target board and redefine yours.
  */
-//#define STEVAL_MKI109V3
-#define NUCLEO_F411RE_X_NUCLEO_IKS01A2
+
+//#define STEVAL_MKI109V3  /* little endian */
+//#define NUCLEO_F411RE    /* little endian */
+//#define SPC584B_DIS      /* big endian */
+
+/* ATTENTION: By default the driver is little endian. If you need switch
+ *            to big endian please see "Endianness definitions" in the
+ *            header file of the driver (_reg.h).
+ */
 
 #if defined(STEVAL_MKI109V3)
 /* MKI109V3: Define communication interface */
 #define SENSOR_BUS hspi2
-
 /* MKI109V3: Vdd and Vddio power supply values */
 #define PWM_3V3 915
 
-#elif defined(NUCLEO_F411RE_X_NUCLEO_IKS01A2)
-/* NUCLEO_F411RE_X_NUCLEO_IKS01A2: Define communication interface */
+#elif defined(NUCLEO_F411RE)
+/* NUCLEO_F411RE: Define communication interface */
 #define SENSOR_BUS hi2c1
+
+#elif defined(SPC584B_DIS)
+/* DISCOVERY_SPC584B: Define communication interface */
+#define SENSOR_BUS I2CD1
 
 #endif
 
 /* Includes ------------------------------------------------------------------*/
 #include <string.h>
 #include <stdio.h>
+#include "lsm6dsr_reg.h"
+
+#if defined(NUCLEO_F411RE)
 #include "stm32f4xx_hal.h"
-#include <lsm6dsr_reg.h>
+#include "usart.h"
 #include "gpio.h"
 #include "i2c.h"
-#if defined(STEVAL_MKI109V3)
+
+#elif defined(STEVAL_MKI109V3)
+#include "stm32f4xx_hal.h"
 #include "usbd_cdc_if.h"
+#include "gpio.h"
 #include "spi.h"
-#elif defined(NUCLEO_F411RE_X_NUCLEO_IKS01A2)
-#include "usart.h"
+#include "tim.h"
+
+#elif defined(SPC584B_DIS)
+#include "components.h"
 #endif
 
 /* Private macro -------------------------------------------------------------*/
+#define    BOOT_TIME            10 //ms
 
 /* Private variables ---------------------------------------------------------*/
 static uint8_t whoamI, rst;
@@ -100,120 +120,68 @@ static int32_t platform_write(void *handle, uint8_t reg,
 static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
                              uint16_t len);
 static void tx_com( uint8_t *tx_buffer, uint16_t len );
+static void platform_delay(uint32_t ms);
 static void platform_init(void);
 
 /* Main Example --------------------------------------------------------------*/
-void example_main_orientation_lsm6dsr(void)
+void lsm6dsr_tilt(void)
 {
   stmdev_ctx_t dev_ctx;
-  /*
-   * Uncomment to configure INT 1
-   */
+  /* Uncomment to configure INT 1 */
   //lsm6dsr_pin_int1_route_t int1_route;
+  /* Uncomment to configure INT 2 */
   lsm6dsr_pin_int2_route_t int2_route;
-  /*
-   *  Initialize mems driver interface.
-   */
+  /* Initialize mems driver interface */
   dev_ctx.write_reg = platform_write;
   dev_ctx.read_reg = platform_read;
-  dev_ctx.handle = &hi2c1;
-  /*
-   * Init test platform.
-   */
+  dev_ctx.handle = &SENSOR_BUS;
+  /* Init test platform */
   platform_init();
-  /*
-   *  Check device ID.
-   */
+  /* Wait sensor boot time */
+  platform_delay(BOOT_TIME);
+  /* Check device ID */
   lsm6dsr_device_id_get(&dev_ctx, &whoamI);
 
   if (whoamI != LSM6DSR_ID)
     while (1);
 
-  /*
-   *  Restore default configuration.
-   */
+  /* Restore default configuration */
   lsm6dsr_reset_set(&dev_ctx, PROPERTY_ENABLE);
 
   do {
     lsm6dsr_reset_get(&dev_ctx, &rst);
   } while (rst);
 
-  /*
-   * Disable I3C interface.
-   */
+  /* Disable I3C interface */
   lsm6dsr_i3c_disable_set(&dev_ctx, LSM6DSR_I3C_DISABLE);
-  /*
-   * Set XL Output Data Rate to 417 Hz.
+  /* Set XL Output Data Rate: The tilt function works at 26 Hz,
+   * so the accelerometer ODR must be set at 26 Hz or higher values
    */
-  lsm6dsr_xl_data_rate_set(&dev_ctx, LSM6DSR_XL_ODR_417Hz);
-  /*
-   * Set 2g full XL scale.
-   */
+  lsm6dsr_xl_data_rate_set(&dev_ctx, LSM6DSR_XL_ODR_26Hz);
+  /* Set 2g full XL scale. */
   lsm6dsr_xl_full_scale_set(&dev_ctx, LSM6DSR_2g);
-  /*
-   * Set threshold to 60 degrees.
-   */
-  lsm6dsr_6d_threshold_set(&dev_ctx, LSM6DSR_DEG_60);
-  /*
-   * LPF2 on 6D/4D function selection.
-   */
-  lsm6dsr_xl_lp2_on_6d_set(&dev_ctx, PROPERTY_ENABLE);
-  /*
-   * To enable 4D mode uncomment next line.
-   * 4D orientation detection disable Z-axis events.
-   */
-  lsm6dsr_4d_mode_set(&dev_ctx, PROPERTY_ENABLE);
-  /*
-   * Uncomment if interrupt generation on Free Fall INT1 pin
-   */
+  /* Enable Tilt in embedded function. */
+  lsm6dsr_tilt_sens_set(&dev_ctx, PROPERTY_ENABLE);
+  /* Uncomment if interrupt generation on Tilt INT1 pin */
   //lsm6dsr_pin_int1_route_get(&dev_ctx, &int1_route);
-  //int1_route.reg.md1_cfg.int1_ff = PROPERTY_ENABLE;
+  //int1_route.emb_func_int1.int1_tilt = PROPERTY_ENABLE;
   //lsm6dsr_pin_int1_route_set(&dev_ctx, &int1_route);
-  /*
-   * Uncomment if interrupt generation on Free Fall INT2 pin
-   */
+  /* Uncomment if interrupt generation on Tilt INT2 pin */
   lsm6dsr_pin_int2_route_get(&dev_ctx, &int2_route);
-  int2_route.md2_cfg.int2_ff = PROPERTY_ENABLE;
+  int2_route.emb_func_int2.int2_tilt = PROPERTY_ENABLE;
   lsm6dsr_pin_int2_route_set(&dev_ctx, &int2_route);
 
-  /*
-   * Wait Events.
-   */
+  /* Uncomment to have interrupt latched */
+  //lsm6dsr_int_notification_set(&dev_ctx, PROPERTY_ENABLE);
+
+  /* Wait Events. */
   while (1) {
-    lsm6dsr_all_sources_t all_source;
-    /*
-     * Check if 6D/4D Orientation events.
-     */
-    lsm6dsr_all_sources_get(&dev_ctx, &all_source);
+    uint8_t is_tilt;
+    /* Check if Tilt events */
+    lsm6dsr_tilt_flag_data_ready_get(&dev_ctx, &is_tilt);
 
-    if (all_source.d6d_src.d6d_ia) {
-      sprintf((char *)tx_buffer, "6D Or. switched to ");
-
-      if (all_source.d6d_src.xh) {
-        strcat((char *)tx_buffer, "XH");
-      }
-
-      if (all_source.d6d_src.xl) {
-        strcat((char *)tx_buffer, "XL");
-      }
-
-      if (all_source.d6d_src.yh) {
-        strcat((char *)tx_buffer, "YH");
-      }
-
-      if (all_source.d6d_src.yl) {
-        strcat((char *)tx_buffer, "YL");
-      }
-
-      if (all_source.d6d_src.zh) {
-        strcat((char *)tx_buffer, "ZH");
-      }
-
-      if (all_source.d6d_src.zl) {
-        strcat((char *)tx_buffer, "ZL");
-      }
-
-      strcat((char *)tx_buffer, "\r\n");
+    if (is_tilt) {
+      sprintf((char *)tx_buffer, "TILT Detected\r\n");
       tx_com(tx_buffer, strlen((char const *)tx_buffer));
     }
   }
@@ -233,20 +201,16 @@ static int32_t platform_write(void *handle, uint8_t reg,
                               uint8_t *bufp,
                               uint16_t len)
 {
-  if (handle == &hi2c1) {
-    HAL_I2C_Mem_Write(handle, LSM6DSR_I2C_ADD_L, reg,
-                      I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
-  }
-
-#ifdef STEVAL_MKI109V3
-
-  else if (handle == &hspi2) {
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
-    HAL_SPI_Transmit(handle, &reg, 1, 1000);
-    HAL_SPI_Transmit(handle, bufp, len, 1000);
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
-  }
-
+#if defined(NUCLEO_F411RE)
+  HAL_I2C_Mem_Write(handle, LSM6DSR_I2C_ADD_L, reg,
+                    I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
+#elif defined(STEVAL_MKI109V3)
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
+  HAL_SPI_Transmit(handle, &reg, 1, 1000);
+  HAL_SPI_Transmit(handle, bufp, len, 1000);
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
+#elif defined(SPC584B_DIS)
+  i2c_lld_write(handle,  LSM6DSR_I2C_ADD_L & 0xFE, reg, bufp, len);
 #endif
   return 0;
 }
@@ -264,28 +228,23 @@ static int32_t platform_write(void *handle, uint8_t reg,
 static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
                              uint16_t len)
 {
-  if (handle == &hi2c1) {
-    HAL_I2C_Mem_Read(handle, LSM6DSR_I2C_ADD_L, reg,
-                     I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
-  }
-
-#ifdef STEVAL_MKI109V3
-
-  else if (handle == &hspi2) {
-    /* Read command */
-    reg |= 0x80;
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
-    HAL_SPI_Transmit(handle, &reg, 1, 1000);
-    HAL_SPI_Receive(handle, bufp, len, 1000);
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
-  }
-
+#if defined(NUCLEO_F411RE)
+  HAL_I2C_Mem_Read(handle, LSM6DSR_I2C_ADD_L, reg,
+                   I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
+#elif defined(STEVAL_MKI109V3)
+  reg |= 0x80;
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
+  HAL_SPI_Transmit(handle, &reg, 1, 1000);
+  HAL_SPI_Receive(handle, bufp, len, 1000);
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
+#elif defined(SPC584B_DIS)
+  i2c_lld_read(handle, LSM6DSR_I2C_ADD_L & 0xFE, reg, bufp, len);
 #endif
   return 0;
 }
 
 /*
- * @brief  Write generic device register (platform dependent)
+ * @brief  Send buffer to console (platform dependent)
  *
  * @param  tx_buffer     buffer to transmit
  * @param  len           number of byte to send
@@ -293,11 +252,27 @@ static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
  */
 static void tx_com(uint8_t *tx_buffer, uint16_t len)
 {
-#ifdef NUCLEO_F411RE_X_NUCLEO_IKS01A2
+#if defined(NUCLEO_F411RE)
   HAL_UART_Transmit(&huart2, tx_buffer, len, 1000);
-#endif
-#ifdef STEVAL_MKI109V3
+#elif defined(STEVAL_MKI109V3)
   CDC_Transmit_FS(tx_buffer, len);
+#elif defined(SPC584B_DIS)
+  sd_lld_write(&SD2, tx_buffer, len);
+#endif
+}
+
+/*
+ * @brief  platform specific delay (platform dependent)
+ *
+ * @param  ms        delay in ms
+ *
+ */
+static void platform_delay(uint32_t ms)
+{
+#if defined(NUCLEO_F411RE) | defined(STEVAL_MKI109V3)
+  HAL_Delay(ms);
+#elif defined(SPC584B_DIS)
+  osalThreadDelayMilliseconds(ms);
 #endif
 }
 
@@ -306,9 +281,11 @@ static void tx_com(uint8_t *tx_buffer, uint16_t len)
  */
 static void platform_init(void)
 {
-#ifdef STEVAL_MKI109V3
+#if defined(STEVAL_MKI109V3)
   TIM3->CCR1 = PWM_3V3;
   TIM3->CCR2 = PWM_3V3;
+  HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_1);
+  HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_2);
   HAL_Delay(1000);
 #endif
 }
