+++ lsm6dsr_STdC/example/lsm6dsr_wake_up.c
@@ -1,14 +1,13 @@
 /*
  ******************************************************************************
- * @file    free_fall.c
+ * @file    wake_up.c
  * @author  Sensors Software Solution Team
- * @brief   This file show the simplest way to detect free fall event
- *      from sensor.
+ * @brief   This file show how to detect wake-up events from sensor.
  *
  ******************************************************************************
  * @attention
  *
- * <h2><center>&copy; Copyright (c) 2019 STMicroelectronics.
+ * <h2><center>&copy; Copyright (c) 2021 STMicroelectronics.
  * All rights reserved.</center></h2>
  *
  * This software component is licensed by ST under BSD 3-Clause license,
@@ -23,18 +22,20 @@
  * This example was developed using the following STMicroelectronics
  * evaluation boards:
  *
- * - STEVAL_MKI109V3
- * - NUCLEO_F411RE + X_NUCLEO_IKS01A2
- *
- * and STM32CubeMX tool with STM32CubeF4 MCU Package
+ * - STEVAL_MKI109V3 + STEVAL-MKI194V1
+ * - NUCLEO_F411RE + X_STEVAL-MKI194V1
+ * - DISCOVERY_SPC584B + STEVAL-MKI194V1
  *
  * Used interfaces:
  *
  * STEVAL_MKI109V3    - Host side:   USB (Virtual COM)
  *                    - Sensor side: SPI(Default) / I2C(supported)
  *
- * NUCLEO_STM32F411RE + X_NUCLEO_IKS01A2 - Host side: UART(COM) to USB bridge
- *                                       - I2C(Default) / SPI(N/A)
+ * NUCLEO_STM32F411RE - Host side: UART(COM) to USB bridge
+ *                    - I2C(Default) / SPI(supported)
+ *
+ * DISCOVERY_SPC584B  - Host side: UART(COM) to USB bridge
+ *                    - Sensor side: I2C(Default) / SPI(supported)
  *
  * If you need to run this example on a different hardware platform a
  * modification of the functions: `platform_write`, `platform_read`,
@@ -48,37 +49,56 @@
  * If a different hardware is used please comment all
  * following target board and redefine yours.
  */
-//#define STEVAL_MKI109V3
-#define NUCLEO_F411RE_X_NUCLEO_IKS01A2
+
+//#define STEVAL_MKI109V3  /* little endian */
+//#define NUCLEO_F411RE    /* little endian */
+//#define SPC584B_DIS      /* big endian */
+
+/* ATTENTION: By default the driver is little endian. If you need switch
+ *            to big endian please see "Endianness definitions" in the
+ *            header file of the driver (_reg.h).
+ */
 
 #if defined(STEVAL_MKI109V3)
 /* MKI109V3: Define communication interface */
 #define SENSOR_BUS hspi2
-
 /* MKI109V3: Vdd and Vddio power supply values */
 #define PWM_3V3 915
 
-#elif defined(NUCLEO_F411RE_X_NUCLEO_IKS01A2)
-/* NUCLEO_F411RE_X_NUCLEO_IKS01A2: Define communication interface */
+#elif defined(NUCLEO_F411RE)
+/* NUCLEO_F411RE: Define communication interface */
 #define SENSOR_BUS hi2c1
+
+#elif defined(SPC584B_DIS)
+/* DISCOVERY_SPC584B: Define communication interface */
+#define SENSOR_BUS I2CD1
 
 #endif
 
 /* Includes ------------------------------------------------------------------*/
 #include <string.h>
 #include <stdio.h>
+#include "lsm6dsr_reg.h"
+
+#if defined(NUCLEO_F411RE)
 #include "stm32f4xx_hal.h"
-#include <lsm6dsr_reg.h>
+#include "usart.h"
 #include "gpio.h"
 #include "i2c.h"
-#if defined(STEVAL_MKI109V3)
+
+#elif defined(STEVAL_MKI109V3)
+#include "stm32f4xx_hal.h"
 #include "usbd_cdc_if.h"
+#include "gpio.h"
 #include "spi.h"
-#elif defined(NUCLEO_F411RE_X_NUCLEO_IKS01A2)
-#include "usart.h"
+#include "tim.h"
+
+#elif defined(SPC584B_DIS)
+#include "components.h"
 #endif
 
 /* Private macro -------------------------------------------------------------*/
+#define    BOOT_TIME            10 //ms
 
 /* Private variables ---------------------------------------------------------*/
 static uint8_t whoamI, rst;
@@ -100,95 +120,79 @@ static int32_t platform_write(void *handle, uint8_t reg,
 static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
                              uint16_t len);
 static void tx_com( uint8_t *tx_buffer, uint16_t len );
+static void platform_delay(uint32_t ms);
 static void platform_init(void);
 
 /* Main Example --------------------------------------------------------------*/
-void example_main_free_fall_lsm6dsr(void)
+void example_main_wake_up_lsm6dsr(void)
 {
   stmdev_ctx_t dev_ctx;
-  /*
-   * Uncomment to configure INT 1
-   */
+  /* Uncomment to configure INT 1 */
   //lsm6dsr_pin_int1_route_t int1_route;
-  /*
-   * Uncomment to configure INT 2
-   */
-  //lsm6dsr_pin_int2_route_t int2_route;
-  /*
-   *  Initialize mems driver interface
-   */
+  /* Uncomment to configure INT 2 */
+  lsm6dsr_pin_int2_route_t int2_route;
+  /* Initialize mems driver interface */
   dev_ctx.write_reg = platform_write;
   dev_ctx.read_reg = platform_read;
-  dev_ctx.handle = &hi2c1;
-  /*
-   * Init test platform
-   */
+  dev_ctx.handle = &SENSOR_BUS;
+  /* Init test platform */
   platform_init();
-  /*
-   *  Check device ID
-   */
+  /* Wait sensor boot time */
+  platform_delay(BOOT_TIME);
+  /* Check device ID */
   lsm6dsr_device_id_get(&dev_ctx, &whoamI);
 
   if (whoamI != LSM6DSR_ID)
     while (1);
 
-  /*
-   *  Restore default configuration
-   */
+  /* Restore default configuration */
   lsm6dsr_reset_set(&dev_ctx, PROPERTY_ENABLE);
 
   do {
     lsm6dsr_reset_get(&dev_ctx, &rst);
   } while (rst);
 
-  /*
-   * Disable I3C interface
-   */
+  /* Disable I3C interface */
   lsm6dsr_i3c_disable_set(&dev_ctx, LSM6DSR_I3C_DISABLE);
-  /*
-   * Set XL Output Data Rate
-   */
+  /* Set XL Output Data Rate to 416 Hz */
   lsm6dsr_xl_data_rate_set(&dev_ctx, LSM6DSR_XL_ODR_417Hz);
-  /*
-   * Set 2g full XL scale
-   */
+  /* Set 2g full XL scale */
   lsm6dsr_xl_full_scale_set(&dev_ctx, LSM6DSR_2g);
-  /*
-   * Enable LIR
-   */
-  lsm6dsr_int_notification_set(&dev_ctx, LSM6DSR_ALL_INT_LATCHED);
-  /*
-   * Set Free Fall duration to 3 and 6 samples event duration
-   */
-  lsm6dsr_ff_dur_set(&dev_ctx, 0x06);
-  lsm6dsr_ff_threshold_set(&dev_ctx, LSM6DSR_FF_TSH_312mg);
-
-  /*
-   * Uncomment if interrupt generation on Free Fall INT1 pin
-   */
+  /* Apply high-pass digital filter on Wake-Up function */
+  lsm6dsr_xl_hp_path_internal_set(&dev_ctx, LSM6DSR_USE_SLOPE);
+  /* Set Wake-Up threshold: 1 LSb corresponds to FS_XL/2^6 */
+  lsm6dsr_wkup_threshold_set(&dev_ctx, 2);
+  /* Uncomment interrupt generation on Wake-Up INT1 pin */
   //lsm6dsr_pin_int1_route_get(&dev_ctx, &int1_route);
-  //int1_route.reg.md1_cfg.int1_ff = PROPERTY_ENABLE;
+  //int1_route.md1_cfg.int1_wu = PROPERTY_ENABLE;
   //lsm6dsr_pin_int1_route_set(&dev_ctx, &int1_route);
+  /* Enable if interrupt generation on Wake-Up INT2 pin */
+  lsm6dsr_pin_int2_route_get(&dev_ctx, &int2_route);
+  int2_route.md2_cfg.int2_wu = PROPERTY_ENABLE;
+  lsm6dsr_pin_int2_route_set(&dev_ctx, &int2_route);
 
-  /*
-   * Uncomment if interrupt generation on Free Fall INT2 pin
-   */
-  //lsm6dsr_pin_int2_route_get(&dev_ctx, &int2_route);
-  //int2_route.reg.md2_cfg.int2_ff = PROPERTY_ENABLE;
-  //lsm6dsr_pin_int2_route_set(&dev_ctx, &int2_route);
-
-  /*
-   * Wait Events
-   */
+  /* Wait Events */
   while (1) {
     lsm6dsr_all_sources_t all_source;
-    /*
-     * Check if Free Fall events
-     */
+    /* Check if Wake-Up events */
     lsm6dsr_all_sources_get(&dev_ctx, &all_source);
 
-    if (all_source.wake_up_src.ff_ia) {
-      sprintf((char *)tx_buffer, "Free Fall Detected\r\n");
+    if (all_source.wake_up_src.wu_ia) {
+      sprintf((char *)tx_buffer, "Wake-Up event on ");
+
+      if (all_source.wake_up_src.x_wu) {
+        strcat((char *)tx_buffer, "X");
+      }
+
+      if (all_source.wake_up_src.y_wu) {
+        strcat((char *)tx_buffer, "Y");
+      }
+
+      if (all_source.wake_up_src.z_wu) {
+        strcat((char *)tx_buffer, "Z");
+      }
+
+      strcat((char *)tx_buffer, " direction\r\n");
       tx_com(tx_buffer, strlen((char const *)tx_buffer));
     }
   }
@@ -208,20 +212,16 @@ static int32_t platform_write(void *handle, uint8_t reg,
                               uint8_t *bufp,
                               uint16_t len)
 {
-  if (handle == &hi2c1) {
-    HAL_I2C_Mem_Write(handle, LSM6DSR_I2C_ADD_L, reg,
-                      I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
-  }
-
-#ifdef STEVAL_MKI109V3
-
-  else if (handle == &hspi2) {
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
-    HAL_SPI_Transmit(handle, &reg, 1, 1000);
-    HAL_SPI_Transmit(handle, bufp, len, 1000);
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
-  }
-
+#if defined(NUCLEO_F411RE)
+  HAL_I2C_Mem_Write(handle, LSM6DSR_I2C_ADD_L, reg,
+                    I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
+#elif defined(STEVAL_MKI109V3)
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
+  HAL_SPI_Transmit(handle, &reg, 1, 1000);
+  HAL_SPI_Transmit(handle, bufp, len, 1000);
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
+#elif defined(SPC584B_DIS)
+  i2c_lld_write(handle,  LSM6DSR_I2C_ADD_L & 0xFE, reg, bufp, len);
 #endif
   return 0;
 }
@@ -239,28 +239,23 @@ static int32_t platform_write(void *handle, uint8_t reg,
 static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
                              uint16_t len)
 {
-  if (handle == &hi2c1) {
-    HAL_I2C_Mem_Read(handle, LSM6DSR_I2C_ADD_L, reg,
-                     I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
-  }
-
-#ifdef STEVAL_MKI109V3
-
-  else if (handle == &hspi2) {
-    /* Read command */
-    reg |= 0x80;
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
-    HAL_SPI_Transmit(handle, &reg, 1, 1000);
-    HAL_SPI_Receive(handle, bufp, len, 1000);
-    HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
-  }
-
+#if defined(NUCLEO_F411RE)
+  HAL_I2C_Mem_Read(handle, LSM6DSR_I2C_ADD_L, reg,
+                   I2C_MEMADD_SIZE_8BIT, bufp, len, 1000);
+#elif defined(STEVAL_MKI109V3)
+  reg |= 0x80;
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_RESET);
+  HAL_SPI_Transmit(handle, &reg, 1, 1000);
+  HAL_SPI_Receive(handle, bufp, len, 1000);
+  HAL_GPIO_WritePin(CS_up_GPIO_Port, CS_up_Pin, GPIO_PIN_SET);
+#elif defined(SPC584B_DIS)
+  i2c_lld_read(handle, LSM6DSR_I2C_ADD_L & 0xFE, reg, bufp, len);
 #endif
   return 0;
 }
 
 /*
- * @brief  Write generic device register (platform dependent)
+ * @brief  Send buffer to console (platform dependent)
  *
  * @param  tx_buffer     buffer to transmit
  * @param  len           number of byte to send
@@ -268,11 +263,27 @@ static int32_t platform_read(void *handle, uint8_t reg, uint8_t *bufp,
  */
 static void tx_com(uint8_t *tx_buffer, uint16_t len)
 {
-#ifdef NUCLEO_F411RE_X_NUCLEO_IKS01A2
+#if defined(NUCLEO_F411RE)
   HAL_UART_Transmit(&huart2, tx_buffer, len, 1000);
-#endif
-#ifdef STEVAL_MKI109V3
+#elif defined(STEVAL_MKI109V3)
   CDC_Transmit_FS(tx_buffer, len);
+#elif defined(SPC584B_DIS)
+  sd_lld_write(&SD2, tx_buffer, len);
+#endif
+}
+
+/*
+ * @brief  platform specific delay (platform dependent)
+ *
+ * @param  ms        delay in ms
+ *
+ */
+static void platform_delay(uint32_t ms)
+{
+#if defined(NUCLEO_F411RE) | defined(STEVAL_MKI109V3)
+  HAL_Delay(ms);
+#elif defined(SPC584B_DIS)
+  osalThreadDelayMilliseconds(ms);
 #endif
 }
 
@@ -281,9 +292,11 @@ static void tx_com(uint8_t *tx_buffer, uint16_t len)
  */
 static void platform_init(void)
 {
-#ifdef STEVAL_MKI109V3
+#if defined(STEVAL_MKI109V3)
   TIM3->CCR1 = PWM_3V3;
   TIM3->CCR2 = PWM_3V3;
+  HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_1);
+  HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_2);
   HAL_Delay(1000);
 #endif
 }
